+++ src/platform/RendererCandidate.ts
@@ -1,0 +1,4 @@
+export interface RendererCandidate {
+	type: string;
+	options?: any;
+}